+++ packages/stylus/package.js
@@ -12,7 +12,8 @@ Package.registerBuildPlugin({
   npmDependencies: {
     stylus: "https://github.com/meteor/stylus/tarball/bb47a357d132ca843718c63998eb37b90013a449", // fork of 0.54.5
     nib: "1.1.2",
-    "autoprefixer-stylus": "0.9.4"
+    "autoprefixer-stylus": "0.9.4",
+    glob: "7.1.2",
   }
 });
 
@@ -26,6 +27,9 @@ Package.onTest(function (api) {
     'stylus_tests.html',
     'stylus_tests.styl',
     'stylus_tests.import.styl',
-    'stylus_tests.js'
-  ],'client');
+    'stylus_tests.js',
+    'test_files/direct/direct.import.styl',
+    'test_files/indirect/1/indirect1.import.styl',
+    'test_files/indirect/2/indirect2.import.styl',
+  ], 'client');
 });