+++ packages/stylus/plugin/compile-stylus.js
@@ -2,6 +2,7 @@ const stylus = Npm.require('stylus');
 const nib = Npm.require('nib');
 const autoprefixer = Npm.require('autoprefixer-stylus');
 const Future = Npm.require('fibers/future');
+const glob = Npm.require('glob');
 const fs = Plugin.fs;
 const path = Plugin.path;
 
@@ -89,16 +90,21 @@ class StylusCompiler extends MultiFileCachingCompiler {
     }
 
     const importer = {
-      find(importPath, paths) {
+      find(importPath, paths, filename) {
         const parsed = parseImportPath(importPath, paths[paths.length - 1]);
         if (! parsed) { return null; }
 
         if (importPath[0] !== '{') {
           // if it is not a custom syntax path, it could be a lookup in a folder
           for (let i = paths.length - 1; i >= 0; i--) {
-            const joined = path.join(paths[i], importPath);
-            if (statOrNull(joined)) {
-              return [joined];
+            let joined = path.join(paths[i], importPath);
+            // if we ended up with a custom syntax path, let's try without
+            if (joined.startsWith('{}/')) {
+              joined = joined.substr(3);
+            }
+            const resolvedPaths = resolvePath(joined);
+            if (resolvedPaths.length) {
+              return resolvedPaths;
             }
           }
         }
@@ -123,7 +129,22 @@ class StylusCompiler extends MultiFileCachingCompiler {
           return fs.readFileSync(filePath, 'utf8');
         }
 
-        const parsed = parseImportPath(filePath);
+        // The `allFiles` Map references package files using a key that starts
+        // from the root of the package. If `filePath` includes a package
+        // path prefix (like "packages/[package name]"), we'll remove it to
+        // make sure the `filePath` can be properly matched to a key in the
+        // `allFiles` Map.
+        let cleanFilePath = filePath;
+        let packageName = inputFile.getPackageName();
+        if (packageName) {
+          packageName = packageName.replace('local-test:', '');
+          const packagePathPrefix = `packages/${packageName}/`;
+          if (filePath.startsWith(packagePathPrefix)) {
+            cleanFilePath = filePath.replace(packagePathPrefix, '');
+          }
+        }
+
+        const parsed = parseImportPath(cleanFilePath);
         const absolutePath = absoluteImportPath(parsed);
 
         referencedImportPaths.push(absolutePath);
@@ -189,10 +210,10 @@ class StylusCompiler extends MultiFileCachingCompiler {
   }
 }
 
-function statOrNull(path) {
-  try {
-    return fs.statSync(path);
-  } catch (e) {
-    return null;
+function resolvePath(path) {
+  let paths = glob.sync(path);
+  if (paths.length === 0) {
+    paths = glob.sync(`**/${path}`);
   }
+  return paths;
 }